docs(Q27): correct the published reason the half ky layout is not the default
docs/performance.rst and docs/numerics.rst both published the attribution
this row refutes -- that the linked-chain gather stops fusing once the ky
extent is no longer the grid's power of two. Compiling the linked derivative
alone and sweeping Ny shows the lowering flat at every extent, Nyc = 32
included, and the step-graph regression was in the byte metric rather than
in the graph.

Both pages keep the conclusion: the evolved state is still two-sided by
default and no part of the 41.9 per cent is claimed as recovered. What
changes is the reason -- the runtime grid flip and the diagnostics/NetCDF
condensation chain, not the layout's cost.

Also sets the two line-budget baselines to the numbers the architecture
checker measures.

+++ tests/tools/profiling/test_runtime_and_scaling_profile_contracts.py
@@ -379,8 +379,12 @@ def test_compiled_memory_stats_report_the_compilers_buffer_assignment() -> None:
     """
 
     stats: dict[str, int] = {}
+    # An explicit dtype: the CI shards run with JAX_ENABLE_X64, under which an
+    # unannotated zeros() is float64 and the argument is twice this size.
     runtime_kernels._compiled_hlo_text(
-        lambda x: jnp.sum(x * 2.0), jnp.zeros((8, 8)), stats=stats
+        lambda x: jnp.sum(x * 2.0),
+        jnp.zeros((8, 8), dtype=jnp.float32),
+        stats=stats,
     )
 
     assert set(stats) >= {
